+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Added
+
+-   [`GenerateFormUrl()`](./docs/forms-client.md#GenerateFormUrl) method
+-   [`EncryptUserToken()`](./docs/forms-client.md#EncryptUserToken) method
+-   [`DecryptUserToken()`](./docs/forms-client.md#DecryptUserToken) method
+
 ## 1.12.0 (2020-11-12)
 
 ### Added